+++ src/main/java/lonestarrr/arconia/common/core/proxy/ClientProxy.java
@@ -3,20 +3,27 @@ package lonestarrr.arconia.common.core.proxy;
 import lonestarrr.arconia.client.core.handler.BlockEntityRendererHandler;
 import lonestarrr.arconia.client.core.handler.ColorHandler;
 import lonestarrr.arconia.client.effects.PotItemTransfers;
+import lonestarrr.arconia.client.gui.render.RootItemRenderer;
 import lonestarrr.arconia.client.particle.ModParticles;
 import lonestarrr.arconia.client.particle.custom.RainbowParticles;
 import lonestarrr.arconia.common.Arconia;
+import lonestarrr.arconia.common.core.RainbowColor;
 import lonestarrr.arconia.common.item.MagicInABottle;
 import lonestarrr.arconia.common.item.ModItems;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
 import net.minecraft.client.renderer.item.ItemProperties;
 import net.minecraft.resources.ResourceLocation;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import net.neoforged.fml.event.lifecycle.FMLLoadCompleteEvent;
+import net.neoforged.neoforge.client.event.ModelEvent;
 import net.neoforged.neoforge.client.event.RegisterColorHandlersEvent;
 import net.neoforged.neoforge.client.event.RegisterParticleProvidersEvent;
+import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
+import net.neoforged.neoforge.client.extensions.common.RegisterClientExtensionsEvent;
 import net.neoforged.neoforge.common.NeoForge;
+import org.jetbrains.annotations.NotNull;
 
 /**
  * Proxy code inspired by: http://jabelarminecraft.blogspot.com/p/minecraft-modding-organizing-your-proxy.html
@@ -30,6 +37,8 @@ public class ClientProxy implements IProxy {
         modBus.addListener(this::registerItemColors);
         modBus.addListener(BlockEntityRendererHandler::registerBlockEntityRenderers);
         modBus.addListener(this::registerParticleFactories);
+        modBus.addListener(this::registerClientExtensions);
+        modBus.addListener(this::loadAdditionalModels);
 
         IEventBus forgeBus = NeoForge.EVENT_BUS;
         forgeBus.addListener(PotItemTransfers::render);
@@ -42,6 +51,37 @@ public class ClientProxy implements IProxy {
     }
 
     private void loadComplete(FMLLoadCompleteEvent event) {
+    }
+
+    private void loadAdditionalModels(ModelEvent.RegisterAdditional event) {
+        // Tree roots are dynamically rendered, and need a baked (item) model, not associated with any item object, to
+        // represent the actual tree root.
+        // Alternatively, one could use a dummy item as well and forego the custom baked model route.
+        event.register(RootItemRenderer.TREE_ROOT_BASE_MODEL);
+    }
+
+    private void registerClientExtensions(RegisterClientExtensionsEvent event) {
+        // Custom renderer for imbued tree roots
+        event.registerItem(
+                new IClientItemExtensions() {
+
+                    private final BlockEntityWithoutLevelRenderer renderer = new RootItemRenderer();
+
+                    @NotNull
+                    @Override
+                    public BlockEntityWithoutLevelRenderer getCustomRenderer() {
+                        return renderer;
+                    }
+
+                },
+                ModItems.getColoredRoot(RainbowColor.RED).get(),
+                ModItems.getColoredRoot(RainbowColor.ORANGE).get(),
+                ModItems.getColoredRoot(RainbowColor.YELLOW).get(),
+                ModItems.getColoredRoot(RainbowColor.GREEN).get(),
+                ModItems.getColoredRoot(RainbowColor.LIGHT_BLUE).get(),
+                ModItems.getColoredRoot(RainbowColor.BLUE).get(),
+                ModItems.getColoredRoot(RainbowColor.PURPLE).get()
+        );
     }
 
     private void registerItemColors(RegisterColorHandlersEvent.Item event) {